+++ mongoengine/base/fields.py
@@ -42,7 +42,7 @@ class BaseField(object):
     def __init__(self, db_field=None, name=None, required=False, default=None,
                  unique=False, unique_with=None, primary_key=False,
                  validation=None, choices=None, verbose_name=None,
-                 help_text=None, null=False, sparse=False, custom_data=None):
+                 help_text=None, null=False, sparse=False, **metadata):
         """
         :param db_field: The database field to store this field in
             (defaults to the name of the field)
@@ -69,7 +69,7 @@ class BaseField(object):
             then the default value is set
         :param sparse: (optional) `sparse=True` combined with `unique=True` and `required=False`
             means that uniqueness won't be enforced for `None` values
-        :param custom_data: (optional) Custom metadata for this field.
+        :param **metadata: (optional) Custom metadata for this field.
         """
         self.db_field = (db_field or name) if not primary_key else '_id'
 
@@ -88,7 +88,8 @@ class BaseField(object):
         self.null = null
         self.sparse = sparse
         self._owner_document = None
-        self.custom_data = custom_data
+
+        self.__dict__.update(metadata)
 
         # Adjust the appropriate creation counter, and save our local copy.
         if self.db_field == '_id':